rename compiler load module to loader
