+++ src/compiler/mod.rs
@@ -1,4 +1,4 @@
 pub(crate) mod codegen;
-pub(crate) mod load;
+pub(crate) mod loader;
 pub(crate) mod model;
 pub(crate) mod syntax;
